tests/mediacapture-depth/BAK/MediaStream-id-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>getUserMedia() creates a stream with a proper id</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#dom-mediastream-id">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external).
  </li>
  <li>
    When prompted, accept to share your depth stream.
  </li>
</ol>

<div id="log"></div>

<script>

promise_test(t => {
  let allowedCharacters = /^[\u0021\u0023-\u0027\u002A-\u002B\u002D-\u002E\u0030-\u0039\u0041-\u005A\u005E-\u007E]*$/;
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    assert_true(stream.id.length === 36, "the media stream id has 36 characters");
    assert_regexp_match(stream.id, allowedCharacters, "the media stream id uses the set of allowed characters");
  });
}, "Tests that a MediaStream is created by getUserMedia() has a correct id");

</script>
